Rearrange migrator, add safeword handling on states
