+++ src/Modules/Migrators/BaseMigrator.ts
@@ -1,0 +1,4 @@
+export abstract class BaseMigrator {
+    abstract get Version(): string;
+    abstract Migrate(fromVersion: string): void;
+}
