+++ src/Modules/core.ts
@@ -10,6 +10,8 @@ import { CollarModule } from "./collar";
 
 import * as semver from "semver";
 import { StateConfig } from "Settings/Models/states";
+import { BaseMigrator } from "./Migrators/BaseMigrator";
+import { StateMigrator } from "./Migrators/StateMigrator";
 
 // Core Module that can handle basic functionality like server handshakes etc.
 // Maybe can consolidate things like hypnosis/suffocation basic state handling too..
@@ -181,55 +183,18 @@ export class CoreModule extends BaseModule {
         this.CheckForMigrations(previousVersion);
     }
 
+    Migrators: BaseMigrator[] = [new StateMigrator()]
+
     CheckForMigrations(fromVersion: string) {
         if (fromVersion[0] == 'v')
             fromVersion = fromVersion.substring(1);
-        if (fromVersion == "0.2.5" || semver.lt(fromVersion, "0.2.5")) {
-            // Migration to StatesModule
-            console.info("Migrating LSCG Data from 0.2.5 --");
-            console.info(`Previous: ${JSON.stringify(Player.LSCG)}`);
-            let anyImmersive = (Player.LSCG.HypnoModule as any).immersive || (Player.LSCG.InjectorModule as any).immersive;
-            Player.LSCG.StateModule.immersive = anyImmersive;
 
-            // Migrate Hypnosis State
-            if (!!(Player.LSCG.HypnoModule as any).existingEye1Color) {
-                let hypnoState = Player.LSCG.StateModule.states.find(s => s.type == "hypnotized");
-                if (!hypnoState) {
-                    hypnoState = <StateConfig>{
-                        type: "hypnotized",
-                        active: (Player.LSCG.HypnoModule as any).hypnotized,
-                        activatedBy: (Player.LSCG.HypnoModule as any).hypnotizedBy,
-                        activatedAt: (Player.LSCG.HypnoModule as any).activatedAt,
-                        recoveredAt: (Player.LSCG.HypnoModule as any).recoveredAt,
-                        activationCount: Player.LSCG.HypnoModule.stats.hypnotizedCount,
-                        extensions: {}
-                    }
-                    Player.LSCG.StateModule.states.push(hypnoState);
-                }
-                hypnoState.extensions["existingEye1Color"] = (Player.LSCG.HypnoModule as any).existingEye1Color;
-                hypnoState.extensions["existingEye1Name"] = (Player.LSCG.HypnoModule as any).existingEye1Name;
-                hypnoState.extensions["existingEye2Color"] = (Player.LSCG.HypnoModule as any).existingEye2Color;
-                hypnoState.extensions["existingEye2Name"] = (Player.LSCG.HypnoModule as any).existingEye2Name;
-                hypnoState.extensions["existingEyeExpression"] = (Player.LSCG.HypnoModule as any).existingEyeExpression;
-            }
-            delete (Player.LSCG.HypnoModule as any).existingEye1Color;
-            delete (Player.LSCG.HypnoModule as any).existingEye1Name;
-            delete (Player.LSCG.HypnoModule as any).existingEye2Color;
-            delete (Player.LSCG.HypnoModule as any).existingEye2Name;
-            delete (Player.LSCG.HypnoModule as any).existingEyeExpression;
-            delete (Player.LSCG.HypnoModule as any).hypnotized;
-            delete (Player.LSCG.HypnoModule as any).hypnotizedBy;
-            delete (Player.LSCG.HypnoModule as any).activatedAt;
-            delete (Player.LSCG.HypnoModule as any).recoveredAt;
-            delete (Player.LSCG.HypnoModule as any).immersive;
+        this.Migrators.forEach(m => {
+            if (semver.lt(fromVersion, m.Version))
+                m.Migrate(fromVersion);
+        });
 
-            // Migrate Sleep State
-            delete (Player.LSCG.InjectorModule as any).immersive;
-            delete (Player.LSCG.InjectorModule as any).brainwashed;
-            delete (Player.LSCG.InjectorModule as any).asleep;
-
-            console.info(`Migrated: ${JSON.stringify(Player.LSCG)}`);
-        }
+        settingsSave();
     }
 
     SendPublicPacket(replyRequested: boolean, type: LSCGMessageModelType = "init") {
